refactor(threads): rename threads to thread
