+++ src/pamiq_core/thread/__init__.py
