+++ src/pamiq_core/thread/base.py
