+++ src/pamiq_core/thread/control_thread.py
