+++ src/pamiq_core/thread/inference_thread.py
