+++ src/pamiq_core/thread/thread_types.py
