+++ src/pamiq_core/thread/training_thread.py
