+++ tests/pamiq_core/interaction/test_agent.py
@@ -8,7 +8,7 @@ from pamiq_core.interaction.agent import Agent
 from pamiq_core.interaction.event_mixin import InteractionEventMixin
 from pamiq_core.model import InferenceModel, InferenceModelsDict
 from pamiq_core.state_persistence import PersistentStateMixin
-from pamiq_core.threads import ThreadEventMixin
+from pamiq_core.thread import ThreadEventMixin
 
 
 class AgentImpl(Agent[str, int]):
